requirements.txt
discord~=1.7.3
python-dotenv~=0.19.1
requests~=2.26.0
Pillow~=8.4.0